# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Furniture ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with mobile + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories with subcategories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # PRODUCTS
...
  /inventory/logs:
    get:
      operationId: listInventoryLogs
      tags: [inventory]
      summary: List inventory logs
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [in, out, adjustment]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Inventory logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLogListResponse"
  /inventory/adjust:
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Manual stock adjustment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustInventoryBody"
      responses:
        "200":
          description: Adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryLog"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: orderId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # PAYMENTS
...
  /manufacturers:
    get:
      operationId: listManufacturers
      tags: [manufacturers]
      summary: List manufacturers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Manufacturers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManufacturerListResponse"

    post:
      operationId: createManufacturer
      tags: [manufacturers]
      summary: Create manufacturer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManufacturerBody"
      responses:
        "201":
          description: Manufacturer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"
  /manufacturers/{id}:
    get:
      operationId: getManufacturer
      tags: [manufacturers]
      summary: Get manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Manufacturer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"

    put:
      operationId: updateManufacturer
      tags: [manufacturers]
      summary: Update manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManufacturerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manufacturer"

    delete:
      operationId: deleteManufacturer
      tags: [manufacturers]
      summary: Delete manufacturer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # PURCHASE ORDERS
...
  /dashboard/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [dashboard]
      summary: Recent orders
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/sales-report:
    get:
      operationId: getSalesReport
      tags: [dashboard]
      summary: Sales report by month
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesReportItem"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # AUTH
    LoginBody:
      type: object
      properties:
        mobile:
          type: string
        password:
          type: string
      required: [mobile, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
          nullable: true
        roleId:
          type: integer
        role:
          $ref: "#/components/schemas/Role"
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, mobile, roleId, isActive, createdAt]
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        permissions:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PermissionSet"
        createdAt:
          type: string
          format: date-time
      required: [id, name, permissions, createdAt]
    PermissionSet:
      type: object
      properties:
        view:
          type: boolean
        add:
          type: boolean
        edit:
          type: boolean
        delete:
          type: boolean
      required: [view, add, edit, delete]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: integer
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: integer
          nullable: true
      required: [name]

    # PRODUCTS
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        price:
          type: number
        gstPercent:
          type: number
        stockQty:
          type: integer
        lowStockThreshold:
          type: integer
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, sku, price, gstPercent, stockQty, lowStockThreshold, createdAt]
...
    InventoryLog:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, productId, type, quantity, createdAt]
    AdjustInventoryBody:
      type: object
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        notes:
          type: string
          nullable: true
      required: [productId, type, quantity]
    InventoryLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ORDERS
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        gstPercent:
          type: number
        totalPrice:
          type: number
      required: [id, productId, quantity, unitPrice, gstPercent, totalPrice]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerMobile:
          type: string
          nullable: true
        customerAddress:
          type: string
          nullable: true
        isGst:
          type: boolean
        customerGstNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, processing, completed, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        taxAmount:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, orderNumber, customerName, isGst, status, items, subtotal, taxAmount, totalAmount, paidAmount, createdAt, updatedAt]
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        orderId:
          type: integer
        order:
          $ref: "#/components/schemas/Order"
          nullable: true
        isGst:
          type: boolean
        cgst:
          type: number
        sgst:
          type: number
        igst:
          type: number
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, invoiceNumber, orderId, isGst, cgst, sgst, igst, totalAmount, createdAt]
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # PAYMENTS
...
    Manufacturer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        specialization:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, isActive, createdAt]
    CreateManufacturerBody:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        specialization:
          type: string
          nullable: true
      required: [name]
    ManufacturerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Manufacturer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # PURCHASE ORDERS
...
    SalesReportItem:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        orderCount:
          type: integer
      required: [month, revenue, orderCount]
...